requirements.txt
python-telegram-bot==20.0
boto3==1.28.0
python-dotenv==1.0.0